- 애니메이션 컨트롤러 제작
- 오브젝트 정보 정의 방식 변경

+++ DX11 Framework/SceneManager.h
@@ -8,7 +8,7 @@ public:
 	CScene				*m_nowScene		= nullptr;
 	CPlayer				*m_pPlayer		= nullptr;
 	CCamera				*m_pCamera		= nullptr;
-	float				fFrameRate		= 0.0f;
+	ULONG				fFrameRate		= 0;
 	float				fTimeElapsed	= 0.0f;
 
 	CSceneManager();
